refactor: centralize function-variant resolution in ir::function_variants
Lift variant-label parsing, dispatch-group collection, and callee
name/function resolution into a single canonical module under src/ir,
re-exported from the ir crate root and consumed by both the codegen
dispatcher and the IR passes. Removes the duplicated copy that lived in
codegen_ir/function_variants.rs (now delegating to the shared module).
